[CDEC-457] Reunite orphans from Pos.Core.Update.VoteId
