+++ core/cardano-sl-core.cabal
@@ -194,7 +194,6 @@ library
                        Pos.Core.Update.SoftwareVersion
                        Pos.Core.Update.SystemTag
                        Pos.Core.Update.Vote
-                       Pos.Core.Update.VoteId
 
   build-depends:       aeson
                      , aeson-options
